% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhIPData-class.R
\name{PhIPData-methods}
\alias{PhIPData-methods}
\alias{counts}
\alias{logfc}
\alias{prob}
\alias{peptideInfo}
\alias{sampleInfo}
\alias{assays<-,PhIPData,list-method}
\alias{assays<-,PhIPData,SimpleList-method}
\alias{counts<-}
\alias{logfc<-}
\alias{prob<-}
\alias{peptideInfo<-}
\alias{sampleInfo<-}
\title{Accessing and Modifying Information in PhIPData objects}
\usage{
counts(x, ...)

logfc(x, ...)

prob(x, ...)

peptideInfo(x, ...)

sampleInfo(x, ...)

\S4method{assays}{PhIPData,list}(x, withDimnames = TRUE, ...) <- value

\S4method{assays}{PhIPData,SimpleList}(x, withDimnames = TRUE, ...) <- value

counts(x) <- value

logfc(x) <- value

prob(x) <- value

peptideInfo(x) <- value

sampleInfo(x) <- value
}
\arguments{
\item{x}{A \code{PhIPData} object}

\item{value}{A \code{matrix}, \code{data.frame}, or \link{Data.Frame} of the
same dimensions (not necessarily the same names)}
}
\value{
Accessors: a \link{DataFrame} object

Setters: a \code{PhIPData} object
}
\description{
Methods to extract and modify \code{assay(s)}(including
convenient functions for \code{counts}, \code{logfc}, and \code{prob}),
\code{sampleInfo}, \code{peptideInfo}, and \code{metadata}.
}
\details{
In addition to the functions detailed in
\linkS4class{RangedSummarizedExperiment}, the \code{PhIPData} class includes
conveniently named functions to quickly access and modify frequently used
components of PhIPData objects.

Replacement functions ensure that names of the replacement object are matched
with the names of the \code{PhIPData} object.

Since packages for identifying differential expression in RNA-seq experiments
are frequently used for estimating fold-changes for peptide enrichments,
the class also includes coercion methods to and from \linkS4class{DGEList}s.
}
\examples{
example("PhIPData")

replacement_dat <- matrix(1L, nrow = 5, ncol = 5)

## SummarizedExperiment Accessors and Setters
assays(phip_obj)
assays(phip_obj)$counts <- replacement_dat
assay(phip_obj, "logfc")
assay(phip_obj, "logfc") <- replacement_dat

## counts
counts(phip_obj)
counts(phip_obj) <- counts_dat

## logfc
logfc(phip_obj)
logfc(phip_obj) <- logfc_dat

## prob
prob(phip_obj)
prob(phip_obj) <- replacement_dat

## coercion functions
as(phip_obj, "DGEList")
as(phip_obj, "List")
as(phip_obj, "list")
}
\seealso{
\code{\link[SummarizedExperiment]{assays}} for
\linkS4class{SummarizedExperiment} operations.
}
