% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhIPData-class.R
\name{PhIPData}
\alias{PhIPData}
\title{PhIPData objects}
\usage{
PhIPData(
  counts = S4Vectors::DataFrame(),
  logfc = S4Vectors::DataFrame(),
  prob = S4Vectors::DataFrame(),
  peptideInfo = S4Vectors::DataFrame(),
  sampleInfo = S4Vectors::DataFrame(),
  metadata = list(),
  .defaultNames = "info"
)
}
\arguments{
\item{counts}{a \code{matrix}, \code{data.frame}, or \code{\linkS4class{DataFrame}}
of integer read counts.}

\item{logfc}{a \code{matrix}, \code{data.frame}, or \code{\linkS4class{DataFrame}}
of log10 estimated fold changes.}

\item{prob}{a \code{matrix}, \code{data.frame}, or \code{\linkS4class{DataFrame}} of
probability values (p-values or posterior probabilities) for enrichment estimates.}

\item{peptideInfo}{a \code{data.frame} or \code{\linkS4class{DataFrame}} of peptide
information. Peptide identifiers (\code{pep_id}) can be specified via row names
of the \code{counts}, \code{logfc}, \code{prob}, or \code{peptideInfo} objects.
Valid construction of a \code{PhIPData} object only requires a \code{pep_id}, but
inclusion of other parameters, such as peptide positions or species, enables use
of other convenient functions for filtering peptides.}

\item{sampleInfo}{a \code{data.frame} or \code{\linkS4class{DataFrame}} of additional sample
information. Unique sample identifiers can be specific via column names of the
\code{counts}, \code{logfc}, \code{prob}, or \code{sampleInfo} objects.}

\item{metadata}{a \code{list} object containing experiment-specific metadata}

\item{.defaultNames}{vector of default names to use when the sample and peptide ID
disagree across the metadata and the \code{counts}, \code{logfc}, and \code{prob} matrices.
If \code{.defaultNames} is of length 1, the same source is used for both peptide and
sample identifiers. If \code{.defaultNames} is longer than 2, the first and second
elements correspond to the default names to use for peptides and samples, respectively.
Valid options are:
\itemize{
    \item{"info": }{names should be taken from the \code{SampleInfo} or \code{peptideInfo} objects.}
    \item{"counts": }{names should be taken from the row/column names of the \code{counts} object.}
    \item{"logfc": }{names should be taken from the row/column names of the \code{logfc} object.}
    \item{"prob": }{names should be taken from the row/column names of the \code{prob} object.}
}}
}
\description{
The \code{PhIPData} class is a matrix-like container designed to organize
results from phage-immunoprecipitation (PhIP-Seq) experiments. Rows in
PhIPData objects represent peptides and columns represent samples. Each
object contains at least three assays:

\itemize{
    \item{\code{counts}:} a matrix of raw read counts,
    \item{\code{logfc}:} a matrix of log10 estimated fold-change in
         comparison to beads-only samples,
    \item{\code{prob}:} a matrix of probabilities associated with whether
         a sample has an enriched antibody response for a peptide.
}

The \code{PhIPData} class extends the \link{RangedSummarizedExperiment} class, so
methods documented in \code{?RangedsummarizedExperiment} and
\code{?SummarizedExperiment} also work on \code{PhIPData} objects.
}
\details{
New \code{PhIPData} objects
can be created using the homonymous constructor.
TODO Edit description. Empty objects are valid objects.
Missing peptideInfo and sampleInfo are also valid objects, empty assays can also be initialized
}
\section{Contructor}{
NA
}

\section{Accessors}{
NA
}

\section{Sy}{
NA
}

