% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{propReads}
\alias{propReads}
\title{Proportion of sample reads}
\usage{
propReads(object, withDimnames = TRUE)
}
\arguments{
\item{object}{\linkS4class{PhIPData} object}

\item{withDimnames}{logical; if true return a matrix with the same dimension
names as the original object.}
}
\value{
A (named) numeric matrix with the same dimensions as the function
input. Matrix values are between 0 and 1.
}
\description{
This function calculates the proportion of total sample reads pulled by
each peptide.
}
\examples{
example("PhIPData")
propReads(phip_obj)

## Return an unnamed matrix
propReads(phip_obj, withDimnames = FALSE)
}
