% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetBeads}
\alias{subsetBeads}
\title{Subset beads-only samples}
\usage{
subsetBeads(object)
}
\arguments{
\item{object}{\linkS4class{PhIPData} object}
}
\value{
a \linkS4class{PhIPData} object.
}
\description{
Function to subset PhIP-seq data for beads-only samples.
}
\examples{
example("PhIPData")
subsetBeads(phip_obj)
}
