% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByPeptideMeta.R
\name{subsetByPeptideMeta}
\alias{subsetByPeptideMeta}
\title{Subset a PhIPData object by virus species.}
\usage{
subsetByPeptideMeta(x, ...)
}
\arguments{
\item{object}{a \code{PhIPData} object}

\item{virus}{regex indicating the virus of interest}
}
\value{
a \code{PhIPData} object
}
\description{
Subset a PhIPData object by virus species.
}
