% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{aliases}
\alias{aliases}
\alias{getAliasPath}
\alias{setAliasPath}
\alias{getAlias}
\alias{setAlias}
\alias{deleteAlias}
\title{Using aliases to subset virus data}
\usage{
getAliasPath()

setAliasPath(path)

getAlias(virus)

setAlias(virus, pattern)

deleteAlias(virus)
}
\arguments{
\item{path}{path to \code{alias.rda}}

\item{virus}{character vector of the alias}

\item{pattern}{character vector of regexpressions corresponding to the alias}
}
\description{
Rather than typing out full viruses names or repeating
regexpressions, users can use aliases as a convenient tool to subset
\code{PhIPData} objects by viral species.
}
\details{
Aliases are saved to an rda file containing only a \code{data.frame}
with two columnes: \code{alias} and \code{pattern}. The \code{alias} column
contains the alias while the \code{pattern} column contains the corresponding
regexpression of interest.

The location of the alias database is returned and defined by
\code{getAliasPath} and \code{setAliasPath}, respectively. By default
\code{getAliasPath} points to the \code{extdata} package folder.

Once an alias is added to the database, it can always be accessed once the
package is loaded. It is recommended to use the functions \code{setAlias}
and \code{deleteAlias} to edit the alias database rather than modify the
.rda file itself. If an alias already exists in the database, \code{setAlias}
replaces the matched pattern.
}
\section{Functions}{
\itemize{
\item \code{getAliasPath}: return the path to the .rda file of aliases.

\item \code{setAliasPath}: set the path to the .rda file of aliases.

\item \code{getAlias}: return a regexpression corresponding to the alias.

\item \code{setAlias}: define/modify the regexpression for an alias.

\item \code{deleteAlias}: remove an alias from the database.
}}

