% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhIPData-class.R
\name{PhIPData}
\alias{PhIPData}
\title{Constructor for PhIPData object.}
\usage{
PhIPData(counts, logfc, prob, peptideInfo, sampleInfo, ...)
}
\arguments{
\item{counts}{matrix of integer read counts}

\item{logfc}{matrix of log10 fold changes}

\item{prob}{matrix of p-values or posterior probabilities for estimated enrichment}

\item{peptideInfo}{data frame of peptide information, must contain columns `pep_id`, `pos_start`, and `pos_end`}

\item{sampleInfo}{data frame of sample information.}
}
\description{
Constructor for PhIPData object.
}
